Extract findInlinableCode()
